sched: fair: avoid little cpus due to sync, prev bias
Important threads can get forced to little cpu's
when the sync or prev_bias hints are followed
blindly. This patch adds a check to see whether
those paths are forcing the task to a cpu that
has less capacity than other cpu's available for
the task. If so, we ignore the sync and prev_bias
and allow the scheduler to make a free decision.

Bug: 117438867
Change-Id: Ie5a99f9a8b65ba9382a8d0de2ae0aad843e558d1
Signed-off-by: Miguel de Dios <migueldedios@google.com>
[
 Also includes:

 sched: fair: Move cpu_is_in_target_set definition

 Causes failures on kernel_test so moving the declaration up.
]
[Lau: replace mid_cap_orig_cpu with max_cap_orig_cpu since 4.9
 doesn't support tri-cluster scheduler ]
Signed-off-by: Lau <laststandrighthere@gmail.com>

Signed-off-by: Lau <laststandrighthere@gmail.com>
Signed-off-by: sushmit1 <sushmit718@gmail.com>