Kprofiles: assign `set_mode` during init
Currently kprofiles the common_notifier_callback() assigns mode the value which is assigned to set_mode,
But, during boot up set_mode is never assigned to anything thus leading to profile being set to 0 during bootup.

To fix this, assign set_mode the value which is being assigned to mode by default, so that common_notifier_callback() can use it to set it properly.

Also, Rename kprofiles_notifier_init() to kprofiles_init() and gaurd set_mode properly to make way for future edits.

Acked-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>